Refactor review option inputs
